24294: utility classes and adapters reworked  - Mox and Mora payloads are no longer generated when  storing in memory  - UUID is no longer created when storing in memory  - Facet and Klasse types are now referenced by identifier (user) rather  than by UUID
