+++ tooling/os2mo_data_import/os2mo_data_import/__init__.py
@@ -1,1 +1,1 @@
-from os2mo_data_import.adapters.builder import Organisation
+from os2mo_data_import.data_types import Organisation
